up1030

+++ src/main/java/org/example/Main.java
@@ -6,8 +6,8 @@ public class Main {
 
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        double data = sc.nextDouble();
+        long data = sc.nextLong();
 
-        System.out.printf("%.11f",data);
+        System.out.printf("%d",data);
     }
 }
